feat: add base controller and remainder update endpoint

+++ src/entities/remainderSchema.js
@@ -20,11 +20,11 @@ export default new EntitySchema({
 			type: "int",
 		},
 		shelf_amount: {
-			type: "numeric",
+			type: "int",
 			default: 0,
 		},
 		order_amount: {
-			type: "numeric",
+			type: "int",
 			default: 0,
 		},
 	},
